refactor: Restructuration modulaire des Edge Functions
Refactorisation de microsoft-auth, microsoft-subscriptions et mailbox-email-processor
suivant l'architecture modulaire de microsoft-webhook.

- ✅ Architecture modulaire avec shared-types.ts
- ✅ Séparation des responsabilités en modules spécialisés
- ✅ Configuration deno.json avec imports mappés
- ✅ Gestion d'erreurs cohérente et logging uniforme
- ✅ Résolution des erreurs TypeScript et lint warnings
- ✅ Code production-ready avec types stricts
