+++ supabase/functions/_shared/encryption.ts
@@ -22,7 +22,7 @@ export interface EncryptedData {
  * Encryption error types
  */
 export class EncryptionError extends Error {
-  constructor(message: string, public cause?: unknown) {
+  constructor(message: string, public override cause?: unknown) {
     super(message)
     this.name = 'EncryptionError'
   }
