+++ supabase/functions/mailbox-email-processor/deno.json
@@ -9,7 +9,8 @@
   },
   "lint": {
     "rules": {
-      "tags": ["recommended"]
+      "tags": ["recommended"],
+      "exclude": ["no-explicit-any"]
     }
   },
   "fmt": {
